+++ src/session/data.rs
@@ -2,7 +2,7 @@
 
 use base64::Engine as _;
 use chrono::{DateTime, Utc};
-use rand::RngCore as _;
+use rand::Rng as _;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;